Updated the Home and Register components with an output value for canceling the register form.
The output value is titled 'cancelRegister', and it is defined inside the Register component.
It is used to emit 'false' so that the Home component can get this event and hide the register form via the showRegister() method.
This is not the optimal way to use child-to-parent communication, in my opinion, but this is what the course wants me to do so far...