-Added a TS service which is used to monitor the count of requests that try to retrieve a page on the website - 'busy requests'.
-Added interceptor which utilizes the busy-service :
This service checks if the request URL exists in the dictionary.
If it does - return the HttpEvent of the URL.
Else - increment the 'busy request' counter, create a delay of 500 ms, add the URL to the dictionary, and decrease the counter.
[^^^This interceptor is meant to simulate the situation of attempting to access multiple uncached URLs.
It is not the final implementation of a production-ready URL cache and/or busy request load balancer.]

-Utilized the interceptor in the nav component - if there is more than 1 busy request, the user will see a loading spinner.